AI-main/requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.3
pydantic-settings==2.1.0
openai==1.11.1
pytest==7.4.0
httpx==0.25.0
python-multipart==0.0.6
gdown==4.7.1
transformers==4.38.2
soundfile==0.12.1
